Change Edit options

+++ src/main/java/com/example/ludogoriesoft/lukeriaerpapi/services/PackageService.java
@@ -47,8 +47,10 @@ public class PackageService {
         Package existingPackage = packageRepository.findByIdAndDeletedFalse(id)
                 .orElseThrow(ChangeSetPersister.NotFoundException::new);
 
-        if (existingPackage.getPhoto() != null) {
+        if (existingPackage.getPhoto() != null && packageDTO.getPhoto() != null) {
             imageService.deleteImageFromSpace(existingPackage.getPhoto());
+        } else {
+            packageDTO.setPhoto(existingPackage.getPhoto());
         }
 
         Package updatedPackage = modelMapper.map(packageDTO, Package.class);
